services/websocket/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
websockets==12.0
pydantic==2.5.0
pydantic-settings==2.1.0
pyjwt==2.8.0
aiokafka==0.10.0
